Update script.js
Changed where it created the button and search bar in the HTML as it kept running and creating new bars when I tested in the console.  Fixed button styling as it wasn't wrapped in an LI tag for the style attributes to take affect. Reworked the querySelector for the active button. Created new list for the search feature and returned it to the page over the regular array.